### 2025-09-07 #8952 @editorialbot
[PRE REVIEW]: Bijx: Bijections and normalizing flows with JAX/NNX
**Submitting author:** @mathisgerdes (<a href="http://orcid.org/0000-0002-2369-1436">Mathis Gerdes</a>)
**Repository:** https://github.com/mathisgerdes/bijx
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.1
