### 2025-09-10 #8952 @crvernon
:wave: @mathisgerdes  - Thank you for your submission to JOSS.  All of our topic editors have full queues right now but I will get you someone assigned as soon as possible.
