### 2025-09-26 464e3c1
Fix issues with GMM implementation.
