### 2025-09-26 d7f9069
Make equivariant convolution and vectorfield work with latest flax version.
