### 2025-09-26 5291ea1
Update to work with flax 0.12.
