### 2025-10-14 211cc5e
Add tests for flax 0.12 pytree compliance & update notebooks.
