### 2025-10-14 69395ab
Migrate to newer nnx.Pytree api, away from flax.struct.dataclass.
