### 2025-10-15 b19bdde
Refactor Crouch-Grossmann integration API: replace boolean is_lie with ManifoldType classes.
Replace the boolean `is_lie` parameter with a more flexible and extensible
`manifold_types` system based on ManifoldType classes. This change improves
code clarity, enables better type safety, and provides a foundation for
adding more sophisticated manifold operations including projections.
