+++ tests/test_solvers.py
@@ -14,6 +14,7 @@ import numpy as np
 import bijx.lie as lie
 
 # Import components to test
+from bijx import cg
 from bijx.cg import CG2, CG3, EULER, crouch_grossmann
 from bijx.solvers import DiffraxConfig, odeint_rk4
 
@@ -62,7 +63,14 @@ class TestCrouchGrossmann:
             # Use fine time step for accuracy
             step_size = 0.01
             x_final = crouch_grossmann(
-                decay_vf, x0, args, t0, t1, step_size, False, tableau
+                decay_vf,
+                x0,
+                args,
+                t0,
+                t1,
+                step_size,
+                manifold_types=cg.Euclidean(),
+                tableau=tableau,
             )
 
             # Compare with exact solution
@@ -91,7 +99,16 @@ class TestCrouchGrossmann:
         t0, t1 = 0.0, 1.0
 
         for tableau in [EULER, CG2, CG3]:
-            y_final = crouch_grossmann(su2_vf, y0, args, t0, t1, 0.05, True, tableau)
+            y_final = crouch_grossmann(
+                su2_vf,
+                y0,
+                args,
+                t0,
+                t1,
+                0.05,
+                manifold_types=cg.Unitary(special=True),
+                tableau=tableau,
+            )
 
             # Exact solution
             y_exact = jax.scipy.linalg.expm(strength * (t1 - t0) * lie.SU2_GEN[2]) @ y0
@@ -128,7 +145,9 @@ class TestCrouchGrossmann:
         omega = 2.0
         args = {"omega": omega}
 
-        r_final = crouch_grossmann(so3_vf, r0, args, 0.0, 1.0, 0.1, True, CG3)
+        r_final = crouch_grossmann(
+            so3_vf, r0, args, 0.0, 1.0, 0.1, manifold_types=cg.Matrix(), tableau=CG3
+        )
 
         # Check orthogonality: R^T R = I
         np.testing.assert_allclose(r_final.T @ r_final, jnp.eye(3), atol=1e-12)
@@ -171,13 +190,20 @@ class TestCrouchGrossmann:
         # Initial state
         state0 = {"position": jnp.array([1.0, 2.0, 0.5]), "rotation": jnp.eye(3)}
 
-        # Geometry specification
-        is_lie = {"position": False, "rotation": True}
+        # Geometry specification - pytree of manifold types
+        manifold_types = {"position": cg.Euclidean(), "rotation": cg.Matrix()}
 
         args = {"damping": 0.5, "omega": 1.0}
 
         state_final = crouch_grossmann(
-            mixed_vf, state0, args, 0.0, 2.0, 0.1, is_lie, CG2
+            mixed_vf,
+            state0,
+            args,
+            0.0,
+            2.0,
+            0.1,
+            manifold_types=manifold_types,
+            tableau=CG2,
         )
 
         # Check Euclidean component decays exponentially
@@ -205,8 +231,8 @@ class TestCrouchGrossmann:
                 0.0,
                 1.0,
                 0.1,
-                True,
-                CG2,
+                manifold_types=cg.Unitary(special=True),
+                tableau=CG2,
             )
             # Simple loss: trace of final state
             return jnp.real(jnp.trace(y_final))
@@ -245,13 +271,34 @@ class TestCrouchGrossmann:
 
         for step_size in step_sizes:
             x_euler = crouch_grossmann(
-                oscillatory_growth_vf, x0, args, t0, t1, step_size, False, EULER
+                oscillatory_growth_vf,
+                x0,
+                args,
+                t0,
+                t1,
+                step_size,
+                manifold_types=cg.Euclidean(),
[…]
