+++ src/bijx/cg.py
@@ -25,6 +25,7 @@ import jax.numpy as jnp
 import numpy as np
 from flax import nnx
 from jax import core, custom_derivatives
+from jax_autovmap import autovmap
 
 
 class ButcherTableau(nnx.Pytree):
@@ -127,71 +128,335 @@ CG3 = ButcherTableau(
 r"""Third-order Crouch-Grossmann method (3rd-order, 3 stages)."""
 
 
-def transport(vect, z, inverse=False):
-    r"""Transport Lie algebra element to tangent space at group element.
+class ManifoldType(nnx.Pytree):
+    r"""Base class defining manifold structure and operations for ODE integration.
 
-    Performs parallel transport of a Lie algebra element (tangent vector
-    at identity) to the tangent space at an arbitrary group element.
+    This abstract class specifies how to perform integration steps on different
+    types of manifolds (Euclidean spaces, Lie groups, etc.). Subclasses define
+    the specific operations needed for their manifold structure.
 
-    For right-invariant vector fields: $X_g = X_e \cdot g$
-    For left-invariant vector fields: $X_g = g \cdot X_e$
+    Key operations:
+        - reduce: Accumulate increments using manifold-appropriate operations
+        - post_stage/post_step: Optional projection back to manifold after updates
+        - Adjoint methods: Handle cotangent space operations for differentiation
 
-    Here, we choose the convention of right-invariant vector fields.
-
-    The transport maps vectors from $T_e G$ (Lie algebra) to $T_g G$.
-
-    Args:
-        vect: Lie algebra element (tangent vector at identity).
-        z: Group element providing the transport destination.
-        inverse: If True, transport from $T_g G$ back to $T_e G$.
-
-    Returns:
-        Transported tangent vector at the specified group element.
-
-    Example:
-        >>> # Transport SU(2) generator to arbitrary group element
-        >>> X = bijx.lie.SU2_GEN[0]  # Generator at identity
-        >>> g = bijx.lie.sample_haar(rng, n=2)  # Arbitrary SU(2) element
-        >>> X_g = transport(X, g)  # Tangent vector at g
-        >>> X_g.shape
-        (2, 2)
+    Common subclasses:
+        - Euclidean: Standard vector spaces with addition
+        - Matrix: Matrix Lie groups using exponential map
+        - Unitary: Unitary groups with optional projection
     """
-    if inverse:
-        z_inv = z.T.conj()
-        return jnp.einsum("...ij,...jk->...ik", vect, z_inv)
-    return jnp.einsum("...ij,...jk->...ik", vect, z)
+
+    def derivative_type(self):
+        """Return the ManifoldType for tangent/cotangent spaces.
+
+        For most manifolds, derivatives live in a different space than the
+        base manifold. This method specifies the manifold type for these spaces.
+        """
+        raise NotImplementedError()
+
+    def reduce(self, x, *deltas):
+        """Accumulate increments on the manifold.
+
+        Args:
+            x: Base point on manifold.
+            *deltas: Sequence of increments to accumulate.
+
+        Returns:
+            Updated point after accumulating all increments.
+        """
+        return sum(deltas, start=x)
+
+    def post_stage(self, x):
+        """Project state back to manifold after each internal stage.
+
+        Args:
+            x: State after stage update.
+
+        Returns:
+            Projected state (identity by default).
+        """
+        return x
+
+    def post_step(self, x):
+        """Project state back to manifold after each full integration step.
+
+        Args:
+            x: State after full step.
+
+        Returns:
+            Projected state (identity by default).
+        """
+        return x
+
+    def post_adjoint_vjp(self, x, adj_vec, v, adj):
[…]
