### 2025-10-28 bbe7987
Add final kernel init option to conv net.
