### 2025-11-15 6293344
Improvements and clarifications on batch shape handling.
- Add note to distribution that default batch determination assumes
  that final bijection does not change pytree tructure and shapes
- Make it easy to override get_batch_shape in subclasses of TRansformed
  for cases where this assumption is violated.
- Add "canonical" shape to shape utility (batch, *space, channel).
