### 2025-11-15 d818ae6
Relax saveat constraint for solver for flexibility & flax updates.
