+++ src/bijx/bijections/continuous.py
@@ -73,7 +73,7 @@ class ContFlowDiffrax(Bijection):
         self.vf_variables = nnx.data(vf_variables)
         self.vf_meta = nnx.data(vf_meta)
         self.config = config
-        assert config.saveat == diffrax.SaveAt(t1=True), "saveat must be t1=True"
+        # assert config.saveat == diffrax.SaveAt(t1=True), "saveat must be t1=True"
 
     def _vf(self, t, state, args):
         """Vector field wrapper for diffrax integration."""
@@ -128,7 +128,7 @@ class ContFlowDiffrax(Bijection):
         this function. Use :meth:`solve_flow` directly for this.
         """
         sol = self.solve_flow(x, log_density, **kwargs)
-        return jax.tree.map(lambda x: x[0], sol.ys)
+        return jax.tree.map(lambda x: x[-1], sol.ys)
 
     def reverse(self, x, log_density, **kwargs):
         """Solve the ODE flow in reverse and return the final state.
@@ -146,7 +146,7 @@ class ContFlowDiffrax(Bijection):
             t_end=self.config.t_start,
             **kwargs,
         )
-        return jax.tree.map(lambda x: x[0], sol.ys)
+        return jax.tree.map(lambda x: x[-1], sol.ys)
 
 
 class ContFlowRK4(Bijection):
@@ -381,7 +381,7 @@ def _ndim_jacobian(func, event_dim):
 
     @partial(jax.vmap, in_axes=(None, 0), out_axes=(None, -1))
     def _jvp(x, tang):
-        x_flat, _info = info.process_and_flatten(x)
+        x_flat, _, _info = info.process_and_flatten(x)
         _func = partial(func_flat, event_shape=_info.event_shape)
         v, jac = jax.jvp(_func, (x_flat,), (tang,))
         return v.reshape(v.shape[:-1] + _info.event_shape), jac
