### 2025-11-15 #8952 @mathisgerdes
Exciting to see this move up the queue!

Since it's been a while, there have been some changes to the code. Would it make sense to pull them into the `paper` branch and update the version to the latest one (v1.0.1 → v1.1.0)? I didn't want to do that before knowing what the procedure or policy is on that.
