### 2025-11-18 #8952 @mathisgerdes
@nkrusch Thanks for the clarifications. I've merged the latest changes into the paper branch. Given your comments it seems to make sense to work with the latest version of the code.
