### 2025-11-22 5c2d222
Update to flax 0.12.1
This flax update included deprecation of param.value
and changed how dataclass and pytree attributes are declared.
