### 2025-11-22 #8952 @nkrusch
Hello @jodemaey, @mrazomej, @anandojha, @mrbuche, and @pmocz

would you be interested in reviewing this submission for JOSS?
