### 2025-11-28 mathisgerdes/bijx#2 @pmocz
(minor) typos
This looks incredible!

I'm reviewing the repo for JOSS. Please use a tool like `typos` or `codespell` (both pip installable) to fix typos in the code/docs for improved clarity

https://github.com/openjournals/joss-reviews/issues/9521#issuecomment-3587050897
