### 2025-11-28 #9521 @nkrusch
Very efficient - thanks a lot @pmocz!
