+++ paper/paper.bib
@@ -23,7 +23,7 @@
 
 @software{flowjax,
   title = {FlowJAX: Distributions and Normalizing Flows in Jax},
-  author = {Daniel Ward},
+  author = {Daniel Ward and Tennessee Hickling and Matthew Mould and Adrian Seyboldt and Gilead Turok},
   url = {https://github.com/danielward27/flowjax},
   version = {17.2.0},
   year = {2025},
