### 2025-11-30 #9521 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
